カードの時間表示の実装
